+++ x/staking/keeper/hooks.go
@@ -9,71 +9,81 @@ import (
 var _ types.StakingHooks = Keeper{}
 
 // AfterValidatorCreated - call hook if registered
-func (k Keeper) AfterValidatorCreated(ctx sdk.Context, valAddr sdk.ValAddress) {
+func (k Keeper) AfterValidatorCreated(ctx sdk.Context, valAddr sdk.ValAddress) error {
 	if k.hooks != nil {
-		k.hooks.AfterValidatorCreated(ctx, valAddr)
+		return k.hooks.AfterValidatorCreated(ctx, valAddr)
 	}
+	return nil
 }
 
 // BeforeValidatorModified - call hook if registered
-func (k Keeper) BeforeValidatorModified(ctx sdk.Context, valAddr sdk.ValAddress) {
+func (k Keeper) BeforeValidatorModified(ctx sdk.Context, valAddr sdk.ValAddress) error {
 	if k.hooks != nil {
-		k.hooks.BeforeValidatorModified(ctx, valAddr)
+		return k.hooks.BeforeValidatorModified(ctx, valAddr)
 	}
+	return nil
 }
 
 // AfterValidatorRemoved - call hook if registered
-func (k Keeper) AfterValidatorRemoved(ctx sdk.Context, consAddr sdk.ConsAddress, valAddr sdk.ValAddress) {
+func (k Keeper) AfterValidatorRemoved(ctx sdk.Context, consAddr sdk.ConsAddress, valAddr sdk.ValAddress) error {
 	if k.hooks != nil {
-		k.hooks.AfterValidatorRemoved(ctx, consAddr, valAddr)
+		return k.hooks.AfterValidatorRemoved(ctx, consAddr, valAddr)
 	}
+	return nil
 }
 
 // AfterValidatorBonded - call hook if registered
-func (k Keeper) AfterValidatorBonded(ctx sdk.Context, consAddr sdk.ConsAddress, valAddr sdk.ValAddress) {
+func (k Keeper) AfterValidatorBonded(ctx sdk.Context, consAddr sdk.ConsAddress, valAddr sdk.ValAddress) error {
 	if k.hooks != nil {
-		k.hooks.AfterValidatorBonded(ctx, consAddr, valAddr)
+		return k.hooks.AfterValidatorBonded(ctx, consAddr, valAddr)
 	}
+	return nil
 }
 
 // AfterValidatorBeginUnbonding - call hook if registered
-func (k Keeper) AfterValidatorBeginUnbonding(ctx sdk.Context, consAddr sdk.ConsAddress, valAddr sdk.ValAddress) {
+func (k Keeper) AfterValidatorBeginUnbonding(ctx sdk.Context, consAddr sdk.ConsAddress, valAddr sdk.ValAddress) error {
 	if k.hooks != nil {
-		k.hooks.AfterValidatorBeginUnbonding(ctx, consAddr, valAddr)
+		return k.hooks.AfterValidatorBeginUnbonding(ctx, consAddr, valAddr)
 	}
+	return nil
 }
 
 // BeforeDelegationCreated - call hook if registered
-func (k Keeper) BeforeDelegationCreated(ctx sdk.Context, delAddr sdk.AccAddress, valAddr sdk.ValAddress) {
+func (k Keeper) BeforeDelegationCreated(ctx sdk.Context, delAddr sdk.AccAddress, valAddr sdk.ValAddress) error {
 	if k.hooks != nil {
-		k.hooks.BeforeDelegationCreated(ctx, delAddr, valAddr)
+		return k.hooks.BeforeDelegationCreated(ctx, delAddr, valAddr)
 	}
+	return nil
 }
 
 // BeforeDelegationSharesModified - call hook if registered
-func (k Keeper) BeforeDelegationSharesModified(ctx sdk.Context, delAddr sdk.AccAddress, valAddr sdk.ValAddress) {
+func (k Keeper) BeforeDelegationSharesModified(ctx sdk.Context, delAddr sdk.AccAddress, valAddr sdk.ValAddress) error {
 	if k.hooks != nil {
-		k.hooks.BeforeDelegationSharesModified(ctx, delAddr, valAddr)
+		return k.hooks.BeforeDelegationSharesModified(ctx, delAddr, valAddr)
 	}
+	return nil
 }
 
 // BeforeDelegationRemoved - call hook if registered
-func (k Keeper) BeforeDelegationRemoved(ctx sdk.Context, delAddr sdk.AccAddress, valAddr sdk.ValAddress) {
+func (k Keeper) BeforeDelegationRemoved(ctx sdk.Context, delAddr sdk.AccAddress, valAddr sdk.ValAddress) error {
 	if k.hooks != nil {
 		k.hooks.BeforeDelegationRemoved(ctx, delAddr, valAddr)
 	}
+	return nil
 }
 
 // AfterDelegationModified - call hook if registered
-func (k Keeper) AfterDelegationModified(ctx sdk.Context, delAddr sdk.AccAddress, valAddr sdk.ValAddress) {
+func (k Keeper) AfterDelegationModified(ctx sdk.Context, delAddr sdk.AccAddress, valAddr sdk.ValAddress) error {
 	if k.hooks != nil {
-		k.hooks.AfterDelegationModified(ctx, delAddr, valAddr)
+		return k.hooks.AfterDelegationModified(ctx, delAddr, valAddr)
 	}
+	return nil
 }
 
 // BeforeValidatorSlashed - call hook if registered
-func (k Keeper) BeforeValidatorSlashed(ctx sdk.Context, valAddr sdk.ValAddress, fraction sdk.Dec) {
+func (k Keeper) BeforeValidatorSlashed(ctx sdk.Context, valAddr sdk.ValAddress, fraction sdk.Dec) error {
 	if k.hooks != nil {
-		k.hooks.BeforeValidatorSlashed(ctx, valAddr, fraction)
+		return k.hooks.BeforeValidatorSlashed(ctx, valAddr, fraction)
 	}
+	return nil
 }